LibWeb: Split ImageStyleValue out of StyleValue.{h,cpp}
